스크립트 파일 정리
1. Minion.cs, MinionAnimationEvent.cs 삭제
2. Minion.cs, MinionAnimationEvent.cs기능 MinionBehaviour.cs에 추가
3. Turret.cs 삭제
4. Turret.cs 기능 TurretBehaviour.cs에 추가
5. Script 폴더 구조 변경
