+++ Assets/Prefabs/SM_capitol_moba_Blue.prefab
@@ -12,7 +12,6 @@ GameObject:
   - component: {fileID: 6310883796395397970}
   - component: {fileID: 1860324445536313544}
   - component: {fileID: 3628685330871821181}
-  - component: {fileID: 8628600808626947613}
   - component: {fileID: 1337658837572553012}
   m_Layer: 7
   m_Name: SM_capitol_moba_Blue
@@ -115,21 +114,6 @@ CapsuleCollider:
   m_Height: 17
   m_Direction: 1
   m_Center: {x: 0, y: 3.2, z: 0}
---- !u!114 &8628600808626947613
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2596947200077733237}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 37d22990f7838be4095eaba0e4243280, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  hp: 0
-  maxHP: 0
-  HPFilledImage: {fileID: 5445702075917492911}
 --- !u!114 &1337658837572553012
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -142,6 +126,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a1a7c051b9230e54aa762bfc85d447d4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  hp: 0
+  maxHP: 0
+  HPFilledImage: {fileID: 5445702075917492911}
   detectionRange: 8
   target: {fileID: 0}
   isAttack: 0
