+++ Assets/Prefabs/SM_capitol_moba_Red.prefab
@@ -43,7 +43,6 @@ GameObject:
   - component: {fileID: 7796002743988005017}
   - component: {fileID: 1745587759261949033}
   - component: {fileID: 6951010585565477691}
-  - component: {fileID: 6146679536312072817}
   - component: {fileID: 4243730351782104217}
   m_Layer: 6
   m_Name: SM_capitol_moba_Red
@@ -145,21 +144,6 @@ CapsuleCollider:
   m_Height: 17
   m_Direction: 1
   m_Center: {x: 0, y: 3.2, z: 0}
---- !u!114 &6146679536312072817
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7161237787725484302}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 37d22990f7838be4095eaba0e4243280, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  hp: 0
-  maxHP: 0
-  HPFilledImage: {fileID: 6040016852954124640}
 --- !u!114 &4243730351782104217
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -172,6 +156,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a1a7c051b9230e54aa762bfc85d447d4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  hp: 0
+  maxHP: 0
+  HPFilledImage: {fileID: 6040016852954124640}
   detectionRange: 8
   target: {fileID: 0}
   isAttack: 0
