--- Assets/Scripts/Minion/MinionAnimationEvents.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: def75dc5d7ddcb643940aa9fc9305d9d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
