+++ Assets/Scripts/Turret/State.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dd3effcd56144e54c8dcca0eef3393cf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
