+++ Assets/Scripts/Turret/State/TurretAttackState.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0f82726e2e53b06449b6489c572d2b26
+guid: 6b473a65f8108b8498b37eb00580b50d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
