+++ Assets/Scripts/Turret/State/TurretDeadState.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d8a1834c00eb9134f98236f34daba2bf
+guid: 2baf917fab93d8447bdf758f43822c77
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
