--- Assets/Scripts/Turret/Turret.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 37d22990f7838be4095eaba0e4243280
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
