--- Assets/Scripts/Turret/TurretDeadState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d635485ca245b0c41b5d5157b9f994f0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
